Change volume for node_modules to be anonymous
Rationale: This volume is not meant to be used over different
builds of the image. Reusing the existing volume with a new
build with different packages creates a problem where
'npm install' is run during image rebuild with a new
package.json/package-lock.json but when the volume is mounted,
actually the old node_modules contents will be used instead.